UI work and tasks. PLEASE TEST

+++ Plan-EZ/src/events/events.js
@@ -18,6 +18,9 @@ window.selectEvent = function(){
     window.currentEvent = JSON.parse(eventSelect.options[eventSelect.selectedIndex].value)
     var selectedEvent = window.currentEvent;
     window.checkboxes = [];
+    window.taskChecks = [];
+    document.getElementById("tasksDiv").innerHTML ="<h3>Tasks</h3>";
+
     if (selectedEvent) {
         window.name = selectedEvent.name;
         window.month = selectedEvent.month;
@@ -26,8 +29,10 @@ window.selectEvent = function(){
         window.endTimeArray = selectedEvent.endTimes;
         window.host = selectedEvent.host;
         window.attendeesArray = selectedEvent.attendees;
+        window.tasksArray = selectedEvent.tasks;
+        window.twentyFour = selectedEvent.twentyFour;
         if (window.host == window.currentUser){
-            document.getElementById("canAttend").innerText = "You are the host of this event!";
+            document.getElementById("canAttend").innerText = "";
             window.youHost = true;
         }
         else{
@@ -36,9 +41,22 @@ window.selectEvent = function(){
         }
 
         var infoString = "Selected Event: " + window.name +" on " + window.month +" "+ window.day + ".<br> Host: " + window.host;
+        if (window.youHost){
+            infoString += " (You)";
+            document.getElementById("yourUserName").innerHTML = "";
+        }
+        else{
+        }
         var eventInfo = document.getElementById("eventInfo");
         eventInfo.innerHTML = infoString;
-        var timeSlotString = "";
+        var timeSlotString = "Times ";
+        if (window.twentyFour) {
+            timeSlotString += "(24 Hour):<br>";
+        }
+        else{
+            timeSlotString += ":<br>";
+
+        }
         document.getElementById('checkBoxDiv').innerHTML="";
 
         for (i=0; i<window.startTimeArray.length; i++){
@@ -61,7 +79,7 @@ window.selectEvent = function(){
               window.checkboxes.push(box);
           }
           else{
-              document.getElementById("submitButton").disabled = true;
+              //document.getElementById("submitButton").disabled = true;
           }
         }
         var attendeesString = "Attendees: ";
@@ -80,20 +98,34 @@ window.selectEvent = function(){
                 attendeesString+= ", ";
             }
         }
+        if (window.tasksArray) {
+            for (i = 0; i < window.tasksArray.length; i++) {
+                var taskDiv = document.createElement("div");
+                taskDiv.style.display = 'inline';
+                var taskP = document.createElement("p");
+                var taskCheck = document.createElement("input");
+                taskCheck.type = "checkbox";
+                taskCheck.id = "taskBox" + i;
+                window.taskChecks.push(taskCheck);
+                if (window.tasksArray[i].user != ""){
+                    taskCheck.disabled = true;
+                }
+                taskP.innerText = window.tasksArray[i].name + " (" + window.tasksArray[i].user +  ")    ";
+                taskP.appendChild(taskCheck);
+                taskDiv.appendChild(taskP);
+                document.getElementById("tasksDiv").appendChild(taskDiv);
+            }
+        }
         document.getElementById("slots").innerHTML = timeSlotString;
         document.getElementById("attendees").innerHTML = attendeesString;
-        if (youHost) {
-            document.getElementById("submitButton").disabled = true;
-        }
-        else{
-            document.getElementById("submitButton").isDisabled = false;
-        }
+
     }
     else{
         document.getElementById("eventInfo").innerHTML = "";
         document.getElementById("attendees").innerHTML = "";
         document.getElementById("slots").innerHTML = "";
         document.getElementById("checkBoxDiv").innerHTML = "";
+        document.getElementById("tasksDiv").innerHTML = "";
         document.getElementById("submitButton").disabled = true;
 
         infoString = "";
@@ -124,23 +156,57 @@ export const ViewModel = DefineMap.extend({
   // add name to the attendees list
   submit() {
     window.people = (window.people +", "+window.a_name);
+    var anyTimes = false;
+    var anyTasks = false;
+    var alreadyAttending = false;
     var userObj = {
         name: window.currentUser,
         timeslots: []
       }
     for(i = 0; i<window.checkboxes.length; i++){
         if (window.checkboxes[i].checked){
+            anyTimes = true;
             userObj.timeslots.push(window.startTimeArray[i] + "-" +window.endTimeArray[i]);
         }
     }
-    window.currentEvent.attendees.push(userObj);
-    var eventIndex = document.getElementById("eventSelect").selectedIndex -1;
-    window.eventArray.splice(eventIndex, 1);
-    window.eventArray.push(window.currentEvent);
-    localStorage.setItem("events", JSON.stringify(window.eventArray));
-    populateEventBox();
-    document.getElementById("eventSelect").selectedIndex = 0;
-    selectEvent();
+    for(i = 0; i<window.taskChecks.length; i++){
+        if (window.taskChecks[i].checked) {
+            anyTasks = true;
+            window.currentEvent.tasks[i].user = window.currentUser;
+        }
+    }
+    for(i=0; i<window.attendeesArray.length;i++){
+        if (window.attendeesArray[i].name == window.currentUser){
+            alreadyAttending = true;
+        }
+    }
+    if ((youHost&&anyTasks) || ((!youHost)&&!alreadyAttending&&anyTimes) || ((!youHost)&&alreadyAttending&&anyTasks)) {
+        if (anyTimes && !alreadyAttending) {
+            window.currentEvent.attendees.push(userObj);
+        }
+
+        var eventIndex = document.getElementById("eventSelect").selectedIndex - 1;
+        window.eventArray.splice(eventIndex, 1);
+        window.eventArray.push(window.currentEvent);
+        localStorage.setItem("events", JSON.stringify(window.eventArray));
+        populateEventBox();
+        window.alert("Success!");
+        document.getElementById("eventSelect").selectedIndex = 0;
+        selectEvent();
+    }
+    else{
+        if (youHost){
+            window.alert("You have not selected any tasks")
+        }
+        else {
+            if (alreadyAttending){
+                window.alert("You are already attending this event and have not selected any tasks.");
+            }
+            else {
+                window.alert("You have not selected any times.");
+            }
+        }
+    }
 
   }
 });
